+++ src/reader/plugins/mark_as_read.py
@@ -2,7 +2,11 @@
 reader.mark_as_read
 ~~~~~~~~~~~~~~~~~~~
 
-Mark added entries of specific feeds as read if their title matches a regex.
+.. module:: reader
+  :noindex:
+
+Mark added entries of specific feeds as read + unimportant
+if their title matches a regex.
 
 To configure, set the ``make_reader_reserved_name('mark_as_read')``
 (by default, ``.reader.mark_as_read``)
@@ -11,6 +15,10 @@ feed metadata to something like::
     {
         "title": ["first-regex", "second-regex"]
     }
+
+
+.. versionchanged:: 2.4
+    Explicitly mark matching entries as unimportant.
 
 
 .. todo::
@@ -62,7 +70,7 @@ def _mark_as_read(reader, entry, status):
 
     for pattern in patterns or ():
         if re.search(pattern, entry.title):
-            reader.mark_entry_as_read(entry)
+            reader._mark_entry_as_dont_care(entry)
             return